refactor: 판매자 평점 페이지 관련 api는 user domain으로 분리
